Rename *_dns_record[s] to *_dns_record_set[s].
