+++ changelogs/fragments/31-record-set.yml
@@ -1,0 +1,2 @@
+breaking_changes:
+- "The hosttech_dns_record and hosttech_dns_records modules have been renamed to hosttech_dns_record_set and hosttech_dns_record_sets, respectively (https://github.com/ansible-collections/community.dns/pull/31)."
