+++ tests/integration/targets/hetzner/aliases
@@ -1,4 +1,4 @@
 unsupported
-hetzner_dns_record
 hetzner_dns_record_info
+hetzner_dns_record_set
 hetzner_dns_zone_info
