+++ tests/integration/targets/hetzner/tasks/main.yml
@@ -2,7 +2,7 @@
 - name: Basic tests
   block:
     - name: Test gracefull error handling
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: none_existing
         prefix: dns_ansible_collection
         state: present
@@ -34,7 +34,7 @@
       register: ZONES
 
     - name: add record
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone_id: "{{ ZONES.zone_id }}"
         prefix: dns_ansible_collection
         state: present
@@ -78,7 +78,7 @@
       check_mode: yes
 
     - name: add record no change
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: dns_ansible_collection
         state: present
@@ -101,7 +101,7 @@
           - RECORD.diff.after.value | count == 1
 
     - name: modify record change in check_mode
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: dns_ansible_collection
         state: present
@@ -123,7 +123,7 @@
           - RECORD.diff.after.value | count == 1
 
     - name: modify record change
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: dns_ansible_collection
         state: present
@@ -144,7 +144,7 @@
           - RECORD.diff.after.value | count == 1
 
     - name: delete record
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: dns_ansible_collection
         type: CNAME
@@ -160,7 +160,7 @@
           - RECORD is changed
 
     - name: delete record no change
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: dns_ansible_collection
         type: CNAME
@@ -175,7 +175,7 @@
           - RECORD is not changed
 
     - name: add single A record
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone_id: "{{ ZONES.zone_id }}"
         prefix: dns_ansible_collection
         state: present
@@ -187,7 +187,7 @@
       register: RECORD
 
     - name: append A record
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone_id: "{{ ZONES.zone_id }}"
         prefix: dns_ansible_collection
         state: present
@@ -201,7 +201,7 @@
       register: RECORD
 
     - name: append A record
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone_id: "{{ ZONES.zone_id }}"
         prefix: dns_ansible_collection
         state: present
@@ -229,7 +229,7 @@
           - RECORD.set.value | count == 3
 
     - name: delete single A record
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: dns_ansible_collection
         state: present
@@ -254,7 +254,7 @@
           - RECORD.set.value | count == 2
 
     - name: delete all A records
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: dns_ansible_collection
         type: A
@@ -277,7 +277,7 @@
           - RECORD.set == {}
 
     - name: add record
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: website1
         state: present
@@ -287,7 +287,7 @@
         hetzner_token: "{{ hetzner_token }}"
 
     - name: replace record
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: website1
         state: present
@@ -317,7 +317,7 @@
           - RECORD.set.value[0] == '127.0.0.2'
 
     - name: append record website1
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: website1
         state: present
@@ -342,7 +342,7 @@
           - RECORD.set.value | count == 2
 
     - name: replace all
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: website1
         state: present
@@ -366,7 +366,7 @@
           - RECORD.set.value | count == 1
 
     - name: remove all records website1
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: website1
         type: A
@@ -390,7 +390,7 @@
 
   always:
     - name: always delete CNAME record
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: dns_ansible_collection
         type: CNAME
@@ -400,7 +400,7 @@
         hetzner_token: "{{ hetzner_token }}"
 
     - name: always delete A record
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: dns_ansible_collection
         type: A
@@ -410,7 +410,7 @@
         hetzner_token: "{{ hetzner_token }}"
 
     - name: always delete A record
-      hetzner_dns_record:
+      hetzner_dns_record_set:
         zone: "{{ hetzner_test_zone }}"
         prefix: website1
         type: A
