+++ tests/integration/targets/hosttech/aliases
@@ -1,4 +1,4 @@
 unsupported
-hosttech_dns_record
 hosttech_dns_record_info
+hosttech_dns_record_set
 hosttech_dns_zone_info
