+++ tests/integration/targets/hosttech/tasks/username-password.yml
@@ -42,7 +42,7 @@
     hosttech_zone_id: "{{ result.zone_id }}"
 
 - name: (username/password) Ensure that test record is not present
-  hosttech_dns_record:
+  hosttech_dns_record_set:
     state: present
     zone: '{{ test_zone }}'
     record: '{{ test_record }}'
@@ -69,7 +69,7 @@
         - not rec.set
 
   - name: (username/password) Create test record
-    hosttech_dns_record:
+    hosttech_dns_record_set:
       state: present
       zone: '{{ test_zone }}'
       record: '{{ test_record }}'
@@ -84,7 +84,7 @@
       hosttech_password: '{{ hosttech_password }}'
 
   - name: (username/password) Modify test record
-    hosttech_dns_record:
+    hosttech_dns_record_set:
       state: present
       zone: '{{ test_zone }}'
       prefix: '{{ test_record_prefix }}'
@@ -115,7 +115,7 @@
         - rec.set.value | length == 2
 
   - name: (username/password) Delete test record
-    hosttech_dns_record:
+    hosttech_dns_record_set:
       state: absent
       zone_id: '{{ hosttech_zone_id }}'
       record: '{{ rec.set.record }}'
@@ -127,7 +127,7 @@
 
   always:
   - name: (username/password) Ensure that test record is not present
-    hosttech_dns_record:
+    hosttech_dns_record_set:
       state: present
       zone: '{{ test_zone }}'
       record: '{{ test_record }}'
